Move robust_copy method to util file
