+++ nemo/core/connectors/save_restore_connector.py
@@ -28,7 +28,7 @@ from omegaconf import DictConfig, OmegaConf
 from omegaconf.omegaconf import open_dict
 
 from nemo.core import classes as nemo_classes  # to avoid circular import do not import ModelPT directly
-from nemo.lightning.io.artifact.file import robust_copy
+from nemo.utils.file_utils import robust_copy
 from nemo.utils import logging, model_utils
 from nemo.utils.app_state import AppState
 from nemo.utils.get_rank import is_global_rank_zero
